Group skip tests to in separate file
Moved tests
Code clean up
