+++ library/src/test/kotlin/com/chuckerteam/chucker/api/ChuckerInterceptorSkipRequestTest.kt
@@ -4,6 +4,7 @@ import com.chuckerteam.chucker.util.ChuckerInterceptorDelegate
 import com.chuckerteam.chucker.util.ClientFactory
 import com.chuckerteam.chucker.util.NoLoggerRule
 import com.chuckerteam.chucker.util.readByteStringBody
+import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.mockk
 import okhttp3.Headers
@@ -28,6 +29,73 @@ internal class ChuckerInterceptorSkipRequestTest {
 
     @TempDir
     lateinit var tempDir: File
+
+    @ParameterizedTest
+    @EnumSource(value = ClientFactory::class)
+    fun `chucker processes all requests when no skipEndpoints are provided`(factory: ClientFactory) {
+        val chuckerInterceptorWithoutSkipping =
+            ChuckerInterceptorDelegate(
+                cacheDirectoryProvider = { tempDir },
+            )
+        val client = factory.create(chuckerInterceptorWithoutSkipping)
+        executeRequestForPath(client, "/", "Response from /")
+        val transaction = chuckerInterceptorWithoutSkipping.expectTransaction()
+        assertThat(transaction.responseBody).isEqualTo("Response from /")
+
+        executeRequestForPath(client, "/skip/path", "Response from /skip/path")
+        val secondTransaction = chuckerInterceptorWithoutSkipping.expectTransaction()
+        assertThat(secondTransaction.responseBody).isEqualTo("Response from /skip/path")
+    }
+
+    @ParameterizedTest
+    @EnumSource(value = ClientFactory::class)
+    fun `chucker skips requests when skipPaths are provided`(factory: ClientFactory) {
+        val chuckerInterceptorWithSkipping =
+            ChuckerInterceptorDelegate(
+                cacheDirectoryProvider = { tempDir },
+                skipPaths =
+                    listOf(
+                        "",
+                        "    ",
+                        "example",
+                        "www.example.com/skip/path",
+                        "example.com/skip/path",
+                        "90",
+                        "https://example/",
+                        "/skip/path",
+                        "/skip//",
+                        "http://localhost:8080/skip/path/ext",
+                    ),
+            )
+        val client = factory.create(chuckerInterceptorWithSkipping)
+
+        executeRequestForPath(client, "", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "    ", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "www.example.com/skip/path", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "example.com/skip/path", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "90", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "https://example/", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "/skip/path", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "/skip//", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+
+        executeRequestForPath(client, "http://localhost:8080/skip/path/ext", "Hello, world!")
+        chuckerInterceptorWithSkipping.expectNoTransactions()
+    }
 
     @ParameterizedTest
     @EnumSource(value = ClientFactory::class)
@@ -301,6 +369,19 @@ internal class ChuckerInterceptorSkipRequestTest {
                 every { headers } returns Headers.Builder().build()
                 every { body } returns null
             }
+    }
+
+    private fun executeRequestForPath(
+        okHttpClient: OkHttpClient,
+        path: String,
+        responseBody: String,
+    ) {
+        val httpUrl =
+            HttpUrl.Builder().scheme("https").host("testexample.com").addPathSegment(path).build()
+
+        val request = Request.Builder().url(server.url(httpUrl.encodedPath)).build()
+        server.enqueue(MockResponse().setBody(responseBody))
+        okHttpClient.newCall(request).execute().readByteStringBody()
     }
 
     private fun executeRequestForEncodedPath(
